refactor: use core/addressof over utility/addressof
The later is deprecated.
